+++ lib/src/main/java/de/siegmar/fastcsv/writer/CsvWriter.java
@@ -419,21 +419,14 @@ public final class CsvWriter implements Closeable, Flushable {
 
         /// Constructs a [CsvWriter] for the specified OutputStream.
         ///
-        /// This is a convenience method for calling [#build(OutputStream, Charset)]
-        /// with the default charset [StandardCharsets#UTF_8].
-        ///
-        /// This build method wraps the given `outputStream` with an [OutputStreamWriter].
-        /// Both this library's internal buffer and the used [OutputStreamWriter] cause deferred writes to the
-        /// underlying stream. This ensures good performance but also means that **you must call [#flush()] or
-        /// [#close()] to ensure that all data is written to the stream!**
+        /// See [#build(OutputStream, Charset)] for details. This is just a convenience method
+        /// for calling it with `charset` set to [StandardCharsets#UTF_8].
         ///
         /// @param outputStream the OutputStream to write CSV data to.
-        /// @return a new CsvWriter instance - never `null`.
+        /// @return a new CsvWriter instance - never `null`. Remember to close it!
         /// @throws NullPointerException if outputStream is `null`
         /// @see #build(OutputStream, Charset)
         public CsvWriter build(final OutputStream outputStream) {
-            Objects.requireNonNull(outputStream, "outputStream must not be null");
-
             return build(outputStream, StandardCharsets.UTF_8);
         }
 
@@ -441,40 +434,47 @@ public final class CsvWriter implements Closeable, Flushable {
         ///
         /// This build method wraps the given `outputStream` with an [OutputStreamWriter].
         /// Both this library's internal buffer and the used [OutputStreamWriter] cause deferred writes to the
-        /// underlying stream. This ensures good performance but also means that you **must call [#flush()] or
-        /// [#close()] to ensure that all data is written to the stream!**
+        /// underlying stream.
+        /// You typically do not need to wrap the given `outputStream` in a [java.io.BufferedOutputStream].
+        /// This ensures good performance but also means that you **must call [#flush()] or [#close()]**
+        /// to ensure that all data is written to the underlying `outputStream`!
         ///
         /// Use [#build(Path,Charset,OpenOption...)] for optimal performance when writing files!
         ///
         /// @param outputStream the OutputStream to write CSV data to.
         /// @param charset      the character set to be used for writing data to the output stream.
-        /// @return a new CsvWriter instance - never `null`.
+        /// @return a new CsvWriter instance - never `null`. Remember to close it!
         /// @throws NullPointerException if outputStream or charset is `null`
         /// @see #build(OutputStream)
         public CsvWriter build(final OutputStream outputStream, final Charset charset) {
             Objects.requireNonNull(outputStream, "outputStream must not be null");
             Objects.requireNonNull(charset, "charset must not be null");
 
-            return csvWriter(new OutputStreamWriter(outputStream, charset), bufferSize, false, autoFlush);
+            return csvWriter(new OutputStreamWriter(outputStream, charset), bufferSize, autoFlush);
         }
 
         /// Constructs a [CsvWriter] for the specified Writer.
         ///
-        /// This library uses built-in buffering (unless [#bufferSize(int)] is used to disable it) but writes
-        /// its internal buffer to the given `writer` at the end of every record write operation. Therefore,
-        /// you probably want to pass in a [java.io.BufferedWriter] to retain good performance.
+        /// This library uses built-in buffering, unless [#bufferSize(int)] is used to disable it.
+        /// You typically do not need to wrap the given `writer` in an [java.io.BufferedWriter].
+        /// This ensures good performance but also means that you **must call [#flush()] or [#close()]**
+        /// to ensure that all data is written to the underlying `writer`!
+        ///
         /// Use [#build(Path,Charset,OpenOption...)] for optimal performance when writing files!
         ///
         /// @param writer the Writer to use for writing CSV data.
-        /// @return a new CsvWriter instance - never `null`.
+        /// @return a new CsvWriter instance - never `null`. Remember to close it!
         /// @throws NullPointerException if writer is `null`
         public CsvWriter build(final Writer writer) {
             Objects.requireNonNull(writer, "writer must not be null");
 
-            return csvWriter(writer, bufferSize, true, autoFlush);
+            return csvWriter(writer, bufferSize, autoFlush);
         }
 
         /// Constructs a [CsvWriter] for the specified Path.
+        ///
+        /// See [#build(Path,Charset,OpenOption...)] for details. This is just a convenience method
+        /// for calling it with `charset` set to [StandardCharsets#UTF_8].
         ///
         /// @param file        the file to write data to.
         /// @param openOptions options specifying how the file is opened.
@@ -504,40 +504,44 @@ public final class CsvWriter implements Closeable, Flushable {
             Objects.requireNonNull(charset, "charset must not be null");
 
             return csvWriter(new OutputStreamWriter(Files.newOutputStream(file, openOptions),
-                charset), bufferSize, false, autoFlush);
+                charset), bufferSize, autoFlush);
         }
 
-        /// Convenience method to write to the console (standard output).
+        /// Convenience method to write to the console ([Console#writer()]).
         ///
         /// Settings [#bufferSize(int)] and [#autoFlush(boolean)] are ignored.
         /// Data is directly written to standard output and flushed after each record.
         ///
         /// Example use:
-        ///
         /// ```
         /// CsvWriter.builder().toConsole()
         ///     .writeRecord("Hello", "world");
         /// ```
         ///
         /// @return a new CsvWriter instance - never `null`.
-        ///     Calls to [CsvWriter#close()] are ignored, standard out remains open.
+        ///     Calls to [CsvWriter#close()] are ignored, console writer remains open.
         /// @throws IllegalStateException if no console is available
         public CsvWriter toConsole() {
             final Console console = System.console();
             if (console == null) {
                 throw new IllegalStateException("No console available");
             }
-            return csvWriter(console.writer(), 0, false, true);
+            return csvWriter(console.writer(), 0, true);
         }
 
         private CsvWriter csvWriter(final Writer writer, final int bufferSize,
-                                    final boolean autoFlushBuffer, final boolean autoFlushWriter) {
-            final Writable writable = bufferSize > 0
-                ? new FastBufferedWriter(writer, bufferSize, autoFlushBuffer, autoFlushWriter)
-                : new UnbufferedWriter(writer, autoFlushWriter);
-
-            return new CsvWriter(writable,
+                                    final boolean autoFlushWriter) {
+            return new CsvWriter(wrapWriter(writer, bufferSize, autoFlushWriter),
                 fieldSeparator, quoteCharacter, commentCharacter, quoteStrategy, lineDelimiter);
+        }
+
+        private static Writable wrapWriter(final Writer writer, final int bufferSize, final boolean autoFlushWriter) {
+            if (bufferSize == 0) {
+                return new UnbufferedWriter(writer, autoFlushWriter);
+            }
+            return autoFlushWriter
+                ? new AutoflushingFastBufferedWriter(writer, bufferSize)
+                : new FastBufferedWriter(writer, bufferSize);
         }
 
         @Override